requirements.txt
streamlit
google-generativeai>=0.3.0
PyPDF2
pandas
matplotlib
python-dotenv
xlsxwriter
